+++ lib/erl2ex/analyze.ex
@@ -13,6 +13,30 @@ defmodule Erl2ex.Analyze do
   alias Erl2ex.Utils
 
 
+  # These are not allowed as names of functions or variables.
+  # The converter will attempt to rename things that use one of these names.
+  # If an exported function uses one of these names, it will require special
+  # handling in both definition and calling.
+  @elixir_reserved_words [
+    :after,
+    :catch,
+    :do,
+    :else,
+    :end,
+    :false,
+    :fn,
+    :nil,
+    :rescue,
+    :true,
+    :__CALLER__,
+    :__DIR__,
+    :__ENV__,
+    :__MODULE__,
+    :__aliases__,
+    :__block__
+  ] |> Enum.into(MapSet.new)
+
+
   # This is a map of Erlang BIFs to equivalent Elixir Kernel functions.
   @autoimport_map %{
     abs: :abs,
@@ -77,109 +101,128 @@ defmodule Erl2ex.Analyze do
   # rename private functions. Exported functions will not be renamed, and will
   # be defined normally, but calling them (and calling the Kernel functions of
   # the same name) will require full qualification.
-  @elixir_auto_imports %{
-    abs: [1],
-    alias!: [1],
-    and: [2],
-    apply: [2, 3],
-    binary_part: [3],
-    binding: [1],
-    bit_size: [1],
-    byte_size: [1],
-    def: [2],
-    defdelegate: [2],
-    defexception: [1],
-    defimpl: [3],
-    defmacro: [2],
-    defmacrop: [2],
-    "defmodule": [2],
-    defoverridable: [1],
-    defp: [2],
-    defprotocol: [2],
-    defstruct: [1],
-    destructure: [2],
-    div: [2],
-    elem: [2],
-    exit: [1],
-    function_exported?: [3],
-    get_and_update_in: [2, 3],
-    get_in: [2],
-    hd: [1],
-    if: [2],
-    in: [2],
-    inspect: [2],
-    is_atom: [1],
-    is_binary: [1],
-    is_bitstring: [1],
-    is_boolean: [1],
-    is_float: [1],
-    is_function: [1, 2],
-    is_integer: [1],
-    is_list: [1],
-    is_map: [1],
-    is_nil: [1],
-    is_number: [1],
-    is_pid: [1],
-    is_port: [1],
-    is_reference: [1],
-    is_tuple: [1],
-    length: [1],
-    macro_exported?: [3],
-    make_ref: [0],
-    map_size: [1],
-    match?: [2],
-    max: [2],
-    min: [2],
-    node: [0, 1],
-    not: [1],
-    or: [2],
-    put_elem: [3],
-    put_in: [2, 3],
-    raise: [1, 2],
-    rem: [2],
-    reraise: [2, 3],
-    round: [1],
-    self: [0],
-    send: [2],
-    sigil_C: [2],
-    sigil_R: [2],
-    sigil_S: [2],
-    sigil_W: [2],
-    sigil_c: [2],
-    sigil_r: [2],
-    sigil_s: [2],
-    sigil_w: [2],
-    spawn: [1, 3],
-    spawn_link: [1, 3],
-    spawn_monitor: [1, 3],
-    struct: [2],
-    struct!: [2],
-    throw: [1],
-    tl: [1],
-    to_char_list: [1],
-    to_string: [1],
-    trunc: [1],
-    tuple_size: [1],
-    unless: [2],
-    update_in: [2, 3],
-    use: [2],
-    var!: [2],
-  }
+  @elixir_auto_imports [
+    # Kernel functions
+    abs: 1,
+    apply: 2,
+    apply: 3,
+    binary_part: 3,
+    bit_size: 1,
+    byte_size: 1,
+    div: 2,
+    elem: 2,
+    exit: 1,
+    function_exported?: 3,
+    get_and_update_in: 3,
+    get_in: 2,
+    hd: 1,
+    inspect: 2,
+    is_atom: 1,
+    is_binary: 1,
+    is_bitstring: 1,
+    is_boolean: 1,
+    is_float: 1,
+    is_function: 1,
+    is_function: 2,
+    is_integer: 1,
+    is_list: 1,
+    is_map: 1,
+    is_number: 1,
+    is_pid: 1,
+    is_port: 1,
+    is_reference: 1,
+    is_tuple: 1,
+    length: 1,
+    macro_exported?: 3,
+    make_ref: 0,
+    map_size: 1,
+    max: 2,
+    min: 2,
+    node: 0,
+    node: 1,
+    not: 1,
+    put_elem: 3,
+    put_in: 3,
+    rem: 2,
+    round: 1,
+    self: 0,
+    send: 2,
+    spawn: 1,
+    spawn: 3,
+    spawn_link: 1,
+    spawn_link: 3,
+    spawn_monitor: 1,
+    spawn_monitor: 3,
+    struct: 2,
+    struct!: 2,
+    throw: 1,
+    tl: 1,
+    trunc: 1,
+    tuple_size: 1,
+    update_in: 3,
+
+    # Kernel macros
+    alias!: 1,
+    and: 2,
+    binding: 1,
+    def: 2,
+    defdelegate: 2,
+    defexception: 1,
+    defimpl: 3,
+    defmacro: 2,
+    defmacrop: 2,
+    "defmodule": 2,
+    defoverridable: 1,
+    defp: 2,
+    defprotocol: 2,
+    defstruct: 1,
+    destructure: 2,
+    get_and_update_in: 2,
+    if: 2,
+    in: 2,
+    is_nil: 1,
+    match?: 2,
+    or: 2,
+    put_in: 2,
+    raise: 1,
+    raise: 2,
+    reraise: 2,
+    reraise: 3,
+    sigil_C: 2,
+    sigil_R: 2,
+    sigil_S: 2,
+    sigil_W: 2,
+    sigil_c: 2,
+    sigil_r: 2,
+    sigil_s: 2,
+    sigil_w: 2,
+    to_char_list: 1,
+    to_string: 1,
+    unless: 2,
+    update_in: 2,
+    use: 2,
+    var!: 2,
+  ] |> Enum.reduce(%{}, fn({k, v}, m) ->
+    Map.update(m, k, [], fn list -> [v | list] end)
+  end)
 
 
   def module(erl_module, _opts \\ []) do
     analysis = build_base_analysis(erl_module)
+    analysis = Enum.reduce(erl_module.forms, analysis, &collect_attr_info/2)
     analysis = Enum.reduce(erl_module.exports, analysis, &collect_exports/2)
     analysis = Enum.reduce(erl_module.forms, analysis, &collect_imports/2)
     analysis = Enum.reduce(erl_module.forms, analysis, &collect_func_info/2)
     analysis = Enum.reduce(analysis.funcs, analysis, &assign_func_names/2)
     analysis = Enum.reduce(erl_module.type_exports, analysis, &collect_type_exports/2)
-    analysis = Enum.reduce(erl_module.forms, analysis, &collect_attr_info/2)
     analysis = Enum.reduce(erl_module.forms, analysis, &collect_record_info/2)
     analysis = Enum.reduce(erl_module.forms, analysis, &collect_macro_info/2)
     analysis = Enum.reduce(erl_module.specs, analysis, &collect_specs/2)
     analysis
   end
+
+
+  def elixir_reserved_words(), do: @elixir_reserved_words
 
 
   def is_exported?(%AnalyzedModule{funcs: funcs}, name, arity) do
@@ -210,9 +253,11 @@ defmodule Erl2ex.Analyze do
     if Map.has_key?(func_info.arities, arity) do
       mapped_name = func_info.func_name
       cond do
-        not is_legal_function_name(mapped_name) or Utils.is_reserved_word(mapped_name) ->
+        not is_legal_function_name(mapped_name) ->
           {:apply, mapped_name}
-        MapSet.member?(@elixir_special_forms, mapped_name) or Map.has_key?(imports, mapped_name) ->
+        MapSet.member?(@elixir_reserved_words, mapped_name)
+            or MapSet.member?(@elixir_special_forms, mapped_name)
+            or Map.has_key?(imports, mapped_name) ->
           {:qualify, mapped_name}
         true ->
           {:bare, mapped_name}
@@ -230,9 +275,11 @@ defmodule Erl2ex.Analyze do
           end
       end
       cond do
-        not is_legal_function_name(mapped_name) or Utils.is_reserved_word(mapped_name) ->
+        not is_legal_function_name(mapped_name) ->
           {:apply, which_module, mapped_name}
-        MapSet.member?(@elixir_special_forms, mapped_name) or !imported ->
+        MapSet.member?(@elixir_reserved_words, mapped_name)
+            or MapSet.member?(@elixir_special_forms, mapped_name)
+            or !imported ->
           {:qualify, which_module, mapped_name}
         true ->
           {:bare, which_module, mapped_name}
@@ -264,6 +311,11 @@ defmodule Erl2ex.Analyze do
 
   def macro_dispatcher_name(%AnalyzedModule{macro_dispatcher: macro_name}) do
     macro_name
+  end
+
+
+  def func_renamer_name(%AnalyzedModule{func_renamer: func_renamer}) do
+    func_renamer
   end
 
 
@@ -322,6 +374,16 @@ defmodule Erl2ex.Analyze do
   end
 
 
+  def is_deffable_function_name(name) do
+    is_legal_function_name(name) and not MapSet.member?(@elixir_reserved_words, name)
+  end
+
+
+  defp is_legal_function_name(name) do
+    Regex.match?(~r/^[_a-z]\w*$/, Atom.to_string(name))
+  end
+
+
   defp build_base_analysis(erl_module) do
     default_imports = @elixir_auto_imports
       |> Enum.map(fn {name, arities} ->
@@ -338,6 +400,14 @@ defmodule Erl2ex.Analyze do
       used_func_names: default_imports |> Map.keys |> Enum.into(MapSet.new)
     }
   end
+
+
+  defp collect_attr_info(%Erl2ex.ErlAttr{name: name}, analysis) do
+    %AnalyzedModule{analysis |
+      used_attr_names: MapSet.put(analysis.used_attr_names, name)
+    }
+  end
+  defp collect_attr_info(_, analysis), do: analysis
 
 
   defp collect_exports({name, arity}, analysis) do
@@ -379,10 +449,12 @@ defmodule Erl2ex.Analyze do
     func_info = %AnalyzedFunc{func_info |
       arities: Map.put_new(func_info.arities, arity, false)
     }
+    used_attr_names = analysis.used_attr_names
     used_func_names = analysis.used_func_names
+    func_renamer = analysis.func_renamer
     if func_info.func_name == {} do
-      bare_name_ok = is_legal_function_name(name) and
-          not Utils.is_reserved_word(name) and
+      can_be_defd = is_deffable_function_name(name)
+      bare_name_ok = can_be_defd and
           not MapSet.member?(@elixir_special_forms, name) and
           not Map.has_key?(analysis.imports, name)
       if func_info.is_exported or bare_name_ok do
@@ -390,17 +462,18 @@ defmodule Erl2ex.Analyze do
           func_name: name,
         }
         used_func_names = MapSet.put(used_func_names, name)
+        if func_renamer == nil and not can_be_defd do
+          func_renamer = Utils.find_available_name("defrenamed", used_attr_names)
+          used_attr_names = MapSet.put(used_attr_names, func_renamer)
+        end
       end
     end
     %AnalyzedModule{analysis |
       funcs: Map.put(analysis.funcs, name, func_info),
-      used_func_names: used_func_names
+      used_func_names: used_func_names,
+      used_attr_names: used_attr_names,
+      func_renamer: func_renamer,
     }
-  end
-
-
-  defp is_legal_function_name(name) do
-    Regex.match?(~r/^[_a-z]\w*$/, Atom.to_string(name))
   end
 
 
@@ -426,14 +499,6 @@ defmodule Erl2ex.Analyze do
       types: Map.put(analysis.types, name, type_info)
     }
   end
-
-
-  defp collect_attr_info(%Erl2ex.ErlAttr{name: name}, analysis) do
-    %AnalyzedModule{analysis |
-      used_attr_names: MapSet.put(analysis.used_attr_names, name)
-    }
-  end
-  defp collect_attr_info(_, analysis), do: analysis
 
 
   defp collect_record_info(%Erl2ex.ErlRecord{name: name, fields: fields}, analysis) do